+++ packages/react-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @phantom/react-ui
 
+## 2.0.0-beta.0
+
+### Major Changes
+
+- New SDKs chain specific interface, prepare for initial beta release
+
+### Patch Changes
+
+- ef68eaf: Key rotation
+- Updated dependencies [ef68eaf]
+- Updated dependencies
+  - @phantom/client@1.0.0-beta.0
+  - @phantom/browser-sdk@1.0.0-beta.0
+  - @phantom/react-sdk@1.0.0-beta.0
+  - @phantom/parsers@1.0.0-beta.0
+
 ## 1.2.5
 
 ### Patch Changes